complaton on windows: replication

+++ sql/log_event.cc
@@ -1024,7 +1024,7 @@ Log_event* Log_event::read_log_event(const uchar *buf, size_t event_len,
     Notice, a pre-checksum FD version forces alg := BINLOG_CHECKSUM_ALG_UNDEF.
   */
   alg= (event_type != FORMAT_DESCRIPTION_EVENT) ?
-    fdle->used_checksum_alg : get_checksum_alg(buf, event_len);
+    fdle->used_checksum_alg : get_checksum_alg(buf, (ulong)event_len);
   // Emulate the corruption during reading an event
   DBUG_EXECUTE_IF("corrupt_read_log_event_char",
     if (event_type != FORMAT_DESCRIPTION_EVENT)
@@ -1036,7 +1036,7 @@ Log_event* Log_event::read_log_event(const uchar *buf, size_t event_len,
       DBUG_SET("-d,corrupt_read_log_event_char");
     }
   );                                                 
-  if (crc_check && event_checksum_test(const_cast<uchar*>(buf), event_len, alg))
+  if (crc_check && event_checksum_test(const_cast<uchar*>(buf), (ulong)event_len, alg))
   {
 #ifdef MYSQL_CLIENT
     *error= "Event crc check failed! Most likely there is event corruption.";
@@ -1167,29 +1167,29 @@ Log_event *Log_event::read_log_event_no_checksum(
     }
     switch(event_type) {
     case QUERY_EVENT:
-      ev= new Query_log_event(buf, event_len, fdle, QUERY_EVENT);
+      ev= new Query_log_event(buf, (uint)event_len, fdle, QUERY_EVENT);
       break;
     case QUERY_COMPRESSED_EVENT:
-      ev= new Query_compressed_log_event(buf, event_len, fdle,
+      ev= new Query_compressed_log_event(buf, (uint)event_len, fdle,
                                          QUERY_COMPRESSED_EVENT);
       break;
     case ROTATE_EVENT:
-      ev= new Rotate_log_event(buf, event_len, fdle);
+      ev= new Rotate_log_event(buf, (uint)event_len, fdle);
       break;
     case BINLOG_CHECKPOINT_EVENT:
-      ev= new Binlog_checkpoint_log_event(buf, event_len, fdle);
+      ev= new Binlog_checkpoint_log_event(buf, (uint)event_len, fdle);
       break;
     case GTID_EVENT:
-      ev= new Gtid_log_event(buf, event_len, fdle);
+      ev= new Gtid_log_event(buf, (uint)event_len, fdle);
       break;
     case GTID_LIST_EVENT:
-      ev= new Gtid_list_log_event(buf, event_len, fdle);
+      ev= new Gtid_list_log_event(buf, (uint)event_len, fdle);
       break;
     case APPEND_BLOCK_EVENT:
-      ev= new Append_block_log_event(buf, event_len, fdle);
+      ev= new Append_block_log_event(buf, (uint)event_len, fdle);
       break;
     case DELETE_FILE_EVENT:
-      ev= new Delete_file_log_event(buf, event_len, fdle);
+      ev= new Delete_file_log_event(buf, (uint)event_len, fdle);
       break;
     case STOP_EVENT:
       ev= new Stop_log_event(buf, fdle);
@@ -1207,10 +1207,10 @@ Log_event *Log_event::read_log_event_no_checksum(
       ev= new Rand_log_event(buf, fdle);
       break;
     case USER_VAR_EVENT:
-      ev= new User_var_log_event(buf, event_len, fdle);
+      ev= new User_var_log_event(buf, (uint)event_len, fdle);
       break;
     case FORMAT_DESCRIPTION_EVENT:
-      ev= new Format_description_log_event(buf, event_len, fdle);
+      ev= new Format_description_log_event(buf, (uint)event_len, fdle);
       break;
 #if defined(HAVE_REPLICATION) 
     case WRITE_ROWS_EVENT_V1:
@@ -1251,26 +1251,26 @@ Log_event *Log_event::read_log_event_no_checksum(
       break;
 
     case TABLE_MAP_EVENT:
-      ev= new Table_map_log_event(buf, event_len, fdle);
+      ev= new Table_map_log_event(buf, (uint)event_len, fdle);
       break;
 #endif
   case PARTIAL_ROW_DATA_EVENT:
-    ev= new Partial_rows_log_event(buf, event_len, fdle);
+    ev= new Partial_rows_log_event(buf, (uint)event_len, fdle);
     break;
     case BEGIN_LOAD_QUERY_EVENT:
-      ev= new Begin_load_query_log_event(buf, event_len, fdle);
+      ev= new Begin_load_query_log_event(buf, (uint)event_len, fdle);
       break;
     case EXECUTE_LOAD_QUERY_EVENT:
-      ev= new Execute_load_query_log_event(buf, event_len, fdle);
+      ev= new Execute_load_query_log_event(buf, (uint)event_len, fdle);
       break;
     case INCIDENT_EVENT:
-      ev= new Incident_log_event(buf, event_len, fdle);
+      ev= new Incident_log_event(buf, (uint)event_len, fdle);
       break;
     case ANNOTATE_ROWS_EVENT:
-      ev= new Annotate_rows_log_event(buf, event_len, fdle);
+      ev= new Annotate_rows_log_event(buf, (uint)event_len, fdle);
       break;
     case START_ENCRYPTION_EVENT:
-      ev= new Start_encryption_log_event(buf, event_len, fdle);
+      ev= new Start_encryption_log_event(buf, (uint)event_len, fdle);
       break;
     case TRANSACTION_PAYLOAD_EVENT:             // MySQL 8.0
       *error=
